+++ packages/hierarchies/src/hierarchies/HierarchyFiltering.ts
@@ -9,12 +9,22 @@ import { GenericNodeKey, GroupingNodeKey, HierarchyNodeKey, InstancesNodeKey } f
 
 /** @public */
 export interface FilterTargetGroupingNodeInfo {
-  /** Key of the grouping node. */
+  /**
+   * Key of the grouping node.
+   * @deprecated in 1.6. This option is no longer needed.
+   */
   key: GroupingNodeKey;
 
   /**
-   * Depth of the grouping node in the hierarchy.
-   * Generally, it can be retrieved from `parentKeys.length`.
+   * Depth up to which nodes in the hierarchy should be expanded.
+   */
+  depth: number;
+}
+
+/** @public */
+export interface FilteringPathAutoExpandOption {
+  /**
+   * Depth up to which nodes in the hierarchy should be expanded.
    */
   depth: number;
 }
@@ -27,8 +37,9 @@ export interface HierarchyFilteringPathOptions {
    * - If it's `true`, then all nodes up to the filter target will have `autoExpand` flag.
    * - If it's an instance of `FilterTargetGroupingNodeInfo`, then all nodes up to the grouping node that matches this property,
    * will have `autoExpand` flag.
+   * - If it's an instance of `FilteringPathAutoExpandOption`, then all nodes up to and including `depth` will have `autoExpand` flag.
    */
-  autoExpand?: boolean | FilterTargetGroupingNodeInfo;
+  autoExpand?: boolean | FilterTargetGroupingNodeInfo | FilteringPathAutoExpandOption;
 }
 
 /**
@@ -186,7 +197,7 @@ export function createHierarchyFilteringHelper(
         | {
             pathMatcher: (identifier: HierarchyNodeIdentifier) => boolean;
           },
-    ): Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined => {
+    ): NodeProps | undefined => {
       if (!hasFilter) {
         return undefined;
       }
@@ -211,7 +222,7 @@ export function createHierarchyFilteringHelper(
      */
     createChildNodePropsAsync: (props: {
       pathMatcher: (identifier: HierarchyNodeIdentifier) => boolean | Promise<boolean>;
-    }): Promise<Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined> | Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined => {
+    }): Promise<NodeProps | undefined> | NodeProps | undefined => {
       if (!hasFilter) {
         return undefined;
       }
@@ -245,13 +256,16 @@ export function createHierarchyFilteringHelper(
   };
 }
 
+/** @public */
+export type NodeProps = Pick<HierarchyNode, "autoExpand" | "filtering"> & { filtering?: { autoExpandDepth?: number } };
+
 type NormalizedFilteringPath = ReturnType<(typeof HierarchyFilteringPath)["normalize"]>;
 
 class MatchingFilteringPathsReducer {
   private _filteredChildrenIdentifierPaths = new Array<NormalizedFilteringPath>();
   private _isFilterTarget = false;
   private _filterTargetOptions = undefined as HierarchyFilteringPathOptions | undefined;
-  private _needsAutoExpand = false;
+  private _needsAutoExpand: boolean | { depth: number } = false;
 
   public constructor(private _hasFilterTargetAncestor: boolean) {}
 
@@ -262,11 +276,16 @@ class MatchingFilteringPathsReducer {
     } else if (path.length > 1) {
       this._filteredChildrenIdentifierPaths.push({ path: path.slice(1), options });
       if (options?.autoExpand) {
-        this._needsAutoExpand = true;
+        if (
+          !this._needsAutoExpand ||
+          (this._needsAutoExpand !== true && (options.autoExpand === true || this._needsAutoExpand.depth < options.autoExpand.depth))
+        ) {
+          this._needsAutoExpand = options.autoExpand;
+        }
       }
     }
   }
-  public getNodeProps(): Pick<HierarchyNode, "filtering" | "autoExpand"> {
+  public getNodeProps(): NodeProps {
     return {
       ...(this._hasFilterTargetAncestor || this._isFilterTarget || this._filteredChildrenIdentifierPaths.length > 0
         ? {
@@ -274,10 +293,11 @@ class MatchingFilteringPathsReducer {
               ...(this._hasFilterTargetAncestor ? { hasFilterTargetAncestor: true } : undefined),
               ...(this._isFilterTarget ? { isFilterTarget: true, filterTargetOptions: this._filterTargetOptions } : undefined),
               ...(this._filteredChildrenIdentifierPaths.length > 0 ? { filteredChildrenIdentifierPaths: this._filteredChildrenIdentifierPaths } : undefined),
+              ...(this._needsAutoExpand && this._needsAutoExpand !== true ? { autoExpandDepth: this._needsAutoExpand.depth } : undefined),
             },
           }
         : undefined),
-      ...(this._needsAutoExpand ? { autoExpand: true } : undefined),
+      ...(this._needsAutoExpand ? { autoExpand: !!this._needsAutoExpand } : undefined),
     };
   }
 }