+++ packages/hierarchies/src/hierarchies/imodel/FilteringHierarchyDefinition.ts
@@ -146,7 +146,9 @@ export class FilteringHierarchyDefinition implements RxjsHierarchyDefinition {
           return (nodeExtraPropsPossiblyPromise instanceof Promise ? from(nodeExtraPropsPossiblyPromise) : of(nodeExtraPropsPossiblyPromise)).pipe(
             map((nodeExtraProps) => {
               if (nodeExtraProps?.autoExpand) {
-                parsedNode.autoExpand = true;
+                const parentLength = parentNode ? 1 + parentNode.parentKeys.length : 0;
+                parsedNode.autoExpand =
+                  nodeExtraProps.filtering?.autoExpandDepth !== undefined && parentLength >= nodeExtraProps.filtering.autoExpandDepth ? undefined : true;
               }
               if (nodeExtraProps?.filtering) {
                 parsedNode.filtering = nodeExtraProps.filtering;